feat: add list aggregation
Implement the list aggregation known from the dokuwiki data plugin. This
feature is useful when an entire table would be to much boilerplate.
Usually one would select only a few columns.

Enable the option sepbyheader to behave analogous to the data plugin.

Report any options that are not valid for a list aggregation as error.
Show that error during the render phase of a page. That way the error is visible
always when the page is rendered, not only after the initial save when it is
parsed.

See #90 and see SPR-710