feat: rescale imagetags so pixel area ~ count
Rescale the images so that images, that are loaded with the same width are
rescaled to contain the same number of pixels (initial width^2), but keep their
aspect ratio. This must be done in js, because the server does not know the
dimensions of the images used. (espacially since we support external
images)