use generated table alias in joins. fixes #246
For joining the multi table, we used the colref of the refrenced multi
column in creating the alias of the joined table. But when referencing
multi columns from different schemas, the colref can be the same for two
columns. This patch now changes it to make use of the
generateTableAlias() method in our query builder, which makes sure all
aliases are unique.