Add support for optgroup-tags for select-fields
In more complex selects we may want to group options by some criteria.
HTML has the <optgroup>-tag for that purpose.

However in Order to not duplicate code, I've to move the handling of
options to the OptGroup-Class completely.

Known Issues:
-------------
* We may have more than one option with the same value and therefore
more than one option may be marked as selected.