RemoteAPICore: do not rely on timestamps of saved revisions
This fixes the remote API's relying on timestamps of files in the attic
directory. While the timestamp of the current page is always the
revision (and actually how the timestamp of the revision is determined),
timestamps of attic files do not necessarily match the revision they
represent. This should also fix random test failures in the remote api
tests.