fix: fix broken cloud tags of multi-lookup columns
Since we group in the SQL query by tag, cloud tags are always single values.
However if the column we are creating the cloud from, is actually a
multi-column then that single value is still in an array. But since that
array can never contain more than 1 value we can simply take the first
one.

The alternative of changing the type was considered but found not to be
practical, because changing the type of a value/column is not intended.
Recreating a new column and type for that value has been found to be
impractical as well.