Add feature to disable relogin.
DoluWiki originally saves the username and password in the browser
cookie, so that user can be automatially re-logged-in when the PHP
session expires. Leaving any data derived from credentials (password
is encrypted, username is base64-encoded) in browser cookies may not
be desirable. This feature replaces that data with a random string.
This causes relogin to be disabled, and further that rememberme will
not work.