Use OTP for non-sticky login
DokuWiki originally stores username/password in cookie so that user
can be re-logged-in automatically. Although the password is encrypted
(with key known only to DokuWiki), leaving any data derived from
credentials in cookie may not be desirable. This feature uses an OTP
stored in the cookie instead, which allows DokuWiki to decrypt password
in its session storage, but only for non-sticky logins. For sticky
logins, the original DokuWiki mechanism is retained.