fix: Add struct instruction to all pages
There was a bug when using struct on a page together with log and
include plugin. The unintended behavior: the struct data would not be
shown after the first heading of the page, but after the first include
if present, or not at all.
The reason for this behavior was that the log-plugin caused the $ID == $this->lastread
check to fail.

This solution removes this check altogether and leaves it to the
renderer to decide if the data should be rendered or not.

SPR-634