Add action plugin hooks back into the ActionRouter
This is not how integration of plugins would ideally be done in this new
system. Ideally an action plugin would actually implement an instance of
AbstractAction and would just fall into the normal flow of actions here.

However to not break a gazillion of existing plugins, this is just add
the existing two events into the new system through the use of a Plugin
action.

Maybe we could add "new" action plugins later.