beginning of a complete refactoring of the menues
This is an attempt to:

* get rid of the super long, complex functions in in/template.php
* make it easy for plugin authors to add their own items to any
  of our menus, regardless of the used template
* continue the progress to make use of SVG in the menues

This takes a similar approach as my actionrefactor branch. Originially
I thought both refactorings could be done in one, merging the
functionality of DokuWiki actions and the menu items. However I couldn't
make it work. So instead we have two separate but similar things.
Maybe they can converge later on.