added mobile menu
This basically reimplements tpl_action_dropdown but makes use of the
other menus (by setting the context to mobile) and thus still triggers
the appropriate events. This makes #1598 obsolete.

tpl_action_dropdown (and the other tpl_action* methods) still need to be
replaced to make use of the new Menu system.