Fix multiple inputs in one label: it's invalid 🙅
Multiple inputs within a single label in invalid HTML. This bug can not
only occur in multi-fields but also in plugin-provided single fields
which may contain multiple input-fields which are then only combined by
javascript.

Thus, we provide an id which every type can assign to the edit-field
best suited. For multi-types this is usually the new-input-template.

Fixes #292