do not disclose email or IP addresses of users through RSS
The feed used to add real or fake email addresses to items (because some
feed formats require them). The default RSS1.0 format did not show them,
but that changed in the recent update of our FeedCreator classes.

Now all code for that has been removed, always showing an
@undisclosed.example.com email address. Because some feed readers may
use the email address to distinguis authors, the mechanism to prepend
the user name was kept.