Update check supports HTTPS
If SSL is not supported, or HTTPS request failed, a HTTP request will be made to ensure the check going. In the log the request type will be indicated, but on the frontend no additional message would be shown (Though I think it's better to notify admins about non-SSL situations, I refer to WordPress, which just gracefully fallback to HTTP with no alert).