fix: fix regex to return table secedit buttons
Since the hid is optional, it must also be optional in the regex. Also
this commit introduced named capture groups to make it more obvious
which part of the regex captures what.

Also there is now an explicit hid generated for tables, to enable
jumping to the correct section after finishing editing.

This was broken in 2571786c763e04c7abbf27c2245a5720878dc3f1 or #1966
respectively.

Known Issues:
* since both title and hid are optional, a hid may be misinterpreted as
a title if the title is not generated.