pass action name by reference. fixes #2117
This fixes plugins that expect that they can change the global $ACT
instead of $event->data in the ACTION_ACT_PREPROCESS event.

This make the whole route a little bit uncleaner but would increase
backwards compatibility.

I'm not sure how widespread the problem is and if adding this is a good
idea.