feat: canonically provide $ACT to javascript
It would be helpful if there were a reliable way to determine the
current mode in javascript, for example if we only want to execute some
javascript in `show` or `preview` modes. This writes the final mode to
the JSINFO object right before it is written out as header, so we can
hopefully capture the actual mode without an plugin changing it later
on.