try to display an error when installer is run without php
When a server has no PHP support enabled, it will either prompt to
download a PHP file or may display it as HTML instead. When a user tries
to run install.php with a server that does the latter, they will see
some weird garbage with the browser trying to interpret raw PHP as HTML.
This pathc adds a warning at the top, telling the user what is going on.