fix(HTTPClient): add handling for 204: No Content responses
This fixes a bug that would occur if the DokuHTTPClient receives a 204
response, e.g. as a result for a request to delete something. It would
still try to read the body, which fails, thus producing a timeout and
finally throwing an exception.

This fix instructs the HTTPClient to not read the (not existing) body if
it receives a 204.