Fix buttons with invalid empty name attributes
The search buttons from the search form on the top of the page
as well as the search page both produce invalid HTML
due to the name attribute being empty.
As the name is required for each of the PHP functions
producing the buttons and is passed as an empty string,
this makes sure that parameters with empty strings get skipped.