split handler.php into multiple files
For now I left Doku_Handler itself as it were. We will need to keep the
class name around for backwards compatibility but should move the class
itself.

I introduced a new ReWriter Interface to formalize how the various call
writer implementations are accessed.

There are a whole bunch of doc blocks missing.