fix: workaround premature metadata generation on page save 🐛
This fixes a bug where the creator of a new page was not saved to metadata.
That could happen if a plugin calls pageinfo() in the IO_WIKIPAGE_WRITE
event AFTER handler. This creates a new .meta file for the page with the
page's mtime as the creation timestamp. That caused the if clause
changed in this commit to evaluate as false and thus the creator key was
not saved.

This commit now also treats the page as newly created if the created
date in the metadata to the file's current mtime.