✨(dw_locktimer) plugins may reuse to add fields and callbacks
Plugins may want to add further fields to be present when saving drafts.
Plugins may want to execute some js functionality that should be timed to
the saved draft/refreshed lock.

If a plugin does another init() to attach the dw_locktimer to its own editor,
then the default callback would be added a second time, causing unexpected
and undesired behavior.

This includes the changes from the following commits:
6ef45cc1c69591eb7facf381ef4bcf88e3aaa1c0
6ca947f3ad455df4fca1a3076b174b7b2688bd89
0fff419cc95b9783dd33ab02ffb3bd7806d5fcde
87bed8b672a166d073948bcb4ca49aaa81dc880c