Change `const` use to `var` for Safari 9 (on iOS)
Safari 9 in iOS does not support the use of `const` in strict mode. Unfortunately, some of us are stuck with Safari 9 due to not being able to upgrade.

I am a big fan of progressive enhancement, but because all javascript code is packed together and sent off to the client, this small incompatibility breaks all javascript for safari 9 browsers. Switching these two keywords makes everything work again.

There are no other uses of `const` in the codebase as far as I can find. This change has little to no impact on anything else, but slighlty extends the range of supported browsers.