fix #2466. Avoid caching half fetched files
When a remote resource exceeds the fetchsize but the remote server does
not return a Content-Length, we read only the fetchsize amount of bytes
but failed to detect that this was a partial read, thus a partial
resource got cached.

This fix will read fetchsize+1, which will then be correctly determined
as too big and thrown away.