adds pagination in the media gallery.
This should work. I used it on the greebo version though so have to check if it works on this development build also.
Main bug is the Ajax. I hope someone can fix it because I'm scratching my head in how to load the pages inside the content. It works in the legacy media manager though.
My javascript knowledge is severely lacking and need to study it more.