Respect case provided by user for internal links when using headers
On a wiki with "useheading" activated, when an editor references an internal page, the link displayed reuses the casing of the heading. This change respects the first-letter case used by the editor and applies it to the heading used to replace the page name. This avoids having disgraceful uppercase letter in the middle of a sentence without having to repeat the page name in the explicit title component of the link just to change the case of the first letter. Hope this makes sense.