Create separate table files for UTF-8 handling
These will be loaded via include, later on and should be cachable by
Op-Cache.

The formatting has been adjusted to have one entry per line to make
diffing much easier in the future.

For now duplicate keys and commented code from the originals have been
kept. But this should probably be cleaned up in the future.

For now these tables are not used, yet.