Remove Google Lucky for non-existing interwiki links
For compatibility renderer will return string '' of $shortcut and $url instead of NULL when seeing a non-existing interwiki link. In the meantime, media and link output in xhtml renderer is adjusted, to show title text instead when src/href is null. In interwiki case, the title will be the "reference" part of the interwiki link.

This makes it possible to also support no URL cases in `interwiki.conf`. Before it will output a URL as `rawurlencode($reference)`, which doesn't make too much sense since it's encoded. However, I am not sure the use case under the current behavior (no URL, but text as `$reference`).

Docs needs to be added to warn renderer plugin developers of this situation.

This fixes #2588.