use inline template instead of .dist file to populate users.auth.php
users.auth.php is the only file in DokuWiki that is generated by using existing .dist file, instead of inline templates in install.php. This may cause problem when install.php is being run to create animal wiki in farm mode, because in this case DOKU_CONF may have no files at all, so install cannot populate users.auth.php properly. This closes #2845, which may be a result of this behavior - without proper comments at the beginning of the file, once the last authplain user is removed (if users has other auth backend as a proxy of authplain), the whole file will get removed, which causes fatal error.