Fix #1115
without this patch calling _acl_add may result in a corrupted acl.auth.php file.
This is the case when we're adding a config for a user and a scope with an
existing config.

The fix is just to call the _acl_del method before perming the addition.

Note that this _acl_add method is currently called from two places:
- from admin_plugin_acl.handle() which was explicitely calling _acl_del before
- from remote_plugin_acl.addAcl, which can actually result in a corrupted file