Avoid flaky tests by waiting full seconds
The waitForTick() method tried to be clever and return early if a second
had passed since the last call already. Unfortunatel this did not work
reliably in our tests and we often had failures on tests that required
unique revision numbers.

Pull requests for a reliable version welcome. Until the we wait a second
on each call.