Include rev_id in new revision link in notification email
Older versions have a "This is an old revision of the document!" message on the top of the page. By including rev_id, user can always see the version the email is referring to, while knowing if it is the latest version by looking at the wiki page.

A hint about this is also added to email text.

This fixes #2196.