Check our code style for every PR
This excludes the currently failing checks with automatic fixes, as they
are out of scope of this commit. The current goal is automatically
reject pull requests that don't me the standards that the rest of our
code base already follows.

While the current major version of nanasess/setup-php claims to provide
PHP 7.4, there are still problems when trying to run phpcs, see
nanasess/setup-php#29. Therefore we are using PHP 7.3 to run these
checks.