Add feature flag for deferred javascript
This adds a feature flag for the jQuery and main-js requests added in
 #2786 and #2958. This adds only a single feature flag since deferring
jQuery without deferring the main javascript request is likely to cause
errors and confusion.

The feature flag defaults to "on" as this should be unproblematic except
for a few plugins. Also, with this flag being on by default, it should
see more usage and is more likely to uncover existing issues.

This feature flag should be removed once this feature is deemed safe.