Upgrade for jQuery.Cookie to 1.4.1
This fixes a problem with badly encoded cookies that may have been set
by other software on the same domain. Eg. cookies encoded in latin1
instead of UTF-8.

Note: this library is abandoned and superseeded by js-cookie but an
upgrade to that is not straight forward (and the benefits for us are
unclear)