Rough fix for ActiveDirectory caching
This is a very rough fix for (some) of the problems described in #3025.
It should at least ensure that subsequent calls to
authad->retrieveUsers() will return the correct data (or data at all).

The changes in here are as ugly as the original code but should be
simple enough to backport to stable for affected users.

A proper fix would involve setting this whole thing on fire and restart
from the ashes.