suppress errors messages in Mailer for non-admins #3014
When ever a mail is sent to the current user, it is usually the only
receiver. If in that case the mail validation fails, sending the whole
mail will fail which will be signalled by a return false of the
Mailer->send() method anyway. This is already signalled to the user.

The more commoin usecase of a validation error is when other people's
addresses fail to validate. These should never be shown to a user.