refactor page and media resolving, introduce ~ shortcut
This moves the resolve_id() type of functions into their own class
hierarchy.

A new shortcut to be used in links is introduced. The tilde ~ can be
used to reference the current page as a namespace. This is useful to
dynamically create a new namespace from an existing page, effectively
making that page the start page. It allows for a more dynamic growth of
the wiki and makes use of the rarer used "startpage named like the
namespace" method for start pages.

The existing code has not been modified, yet and continues to use the
old, now deprecated methods. Some tests are still failing - before they
are fixed, the expected behaviour needs to be discussed.