first go a refactoring the media manager
For now this only adds code. This takes care of encapsulating media file
information and has some UI classes to display those files in the
fullscreen media manager and the popup.

The code is untested and not integrated to run, yet. Next step is to
replace the parts in the media manager to call this new code (and fix
any bugs).

I tried not to change much HTML yet. This can be done later. For now the
goal is to just make it run again.