test: fix two tests on PHP8
On PHP8 trying to access a non-existing array key leads to a warning,
which we might treat as errors?

This incidentally also fixes another test downstream which broke because
this rendering test changes the global plugin controller, but doesn't
clean it up that if the test errors.

-9000 is used as the position to make it obvious that this number is
meaningless for this test.