test: run test in separate process in case of error
If there is an error when executing the method under test, then this
test caused the downstream test \remoteapicore_test::test_getBacklinks
to unexpectedly fail. Probably, because the global plugin controller
would not have been reset to its original value.

Running this test in its own process ensures that it won't affect any
other test, even if it messes with the global state and doesn't clean up.