Keep toolbar right above the page content text input
In case plugins add content to the edit form they might want to
do so at arbitrary positions (also at the top of the page). In
such case we want to make sure, that the toolbar is always right
above the input field in the form and doesn't get separated by the
content added through the plugin.

Signed-off-by: Frieder Schrempf <dev@fris.de>