development CLI
This CLI is basically a replacement for pluginwiz.dokuwiki.org - It
allows the creation of skeleton files for plugins and templates. Unlike
the wizard, it can also be used to add additional files to existing
plugins.

Skeleton files are downloaded automatically from the Plugin Wizard
repository.

Support for templates may not work 100% yet.